fix swr -> redux (darkmode)
